+++ spec/acceptance/api/v2/appeals/contacts_spec.rb
@@ -58,6 +58,7 @@ resource 'Appeals > Contacts' do
       uncompleted_tasks_count
       updated_at
       updated_in_db_at
+      website
     )
   end