+++ spec/acceptance/api/v2/contacts/merges_spec.rb
@@ -72,6 +72,7 @@ resource 'Contacts > Merges' do
       uncompleted_tasks_count
       updated_at
       updated_in_db_at
+      website
     )
   end