+++ spec/acceptance/api/v2/contacts_spec.rb
@@ -94,6 +94,7 @@ resource 'Contacts' do
       uncompleted_tasks_count
       updated_at
       updated_in_db_at
+      website
     )
   end